vectordb iterator: iterate through all the vectors in the closest <nprobe> centroids
Summary:
The current KNN search only returns k number of records to SQL layer. If there are some additional filters in the query, the server might return less than the expected number of records to the client. Thus, we need a fallback execution plan for those scenarios.

Instead of using Fasis KNN to find the k nearest neighbors, this vector  iterator solution 1) find a list id of the centroids that are closest to the query vector, 2) go through all those centroid one by one, for each centroid, go through all the vectors in the centroid in pk order, and return those records to SQL layer one by one. 3) the results return to SQL layer are sorted by in PK order for each centroid instead of distance order, so when vector iterator search is used, we add filesort in the SQL layer to make sure the results are sorted in order.

The solution consists of the major parts:
1. A session variable to kill switch between KNN search versus the centroid-based iterator.
2. Use the Fasis API to find the centroids that are closest to the query vector.
3. Create a Rdb_vector_iterator based on the list id of the centroids to iterate through all vectors for each centroid.
4. When the session variable is enabled, add filesort to the query to make sure the results are sorted in the SQL layer.

Differential Revision: D56256641

fbshipit-source-id: ea2060539fb04cefe61a2b1ecee9dbfc232a35e1