Revert "innodb: Add thd_wait_begin/end for Sync IO"
Summary:
This reverts commit 721fbd660dad59614e8759b74a200ea7f0fda8f8.

The stall mentioned in the commit above already has DISKIO wait begin/end in `buf_read_page_low`. So this extra DISKIO wait is not needed for that. However, this API is also used to write to the log on the commit code path while the binlog mutex is held, which causes lots of commit stalls.

Differential Revision: D56561026

fbshipit-source-id: 868bc18ceb3a24548c567891a8c50af01c130d3e