Avoid aborting server when purge_logs() clashes with rli_relay_log_raft_reset()
Summary:
Within purge_logs(), after we purge binlogs, it needs to update the binlog's index file. It does this in a crashproof manner within move_crash_safe_index_file_to_index_file() which makes a copy of the index file, modifies it, deletes original index file and renames copied index file in place of the original index file. We abort the server when the rename fails. The rename fails because rli_relay_log_raft_reset() can be running, which restores the index file.

To avoid this, rli_relay_log_raft_reset() should acquire the global mysql_bin_log's index lock.

Since rli_relay_log_raft_reset() is called when we transition from primary to secondary, I feel it is OK to take a hit on grabbing the index lock.

Differential Revision: D57198631

fbshipit-source-id: 88a08f1658b736fcf4af41e2c215e858f1d1d9b3