Register file on noop
Summary:
When Raft becomes a leader we first append a no-op to the relay log.
Then we catch up to this no-op before switching to binlog mode.

The thing is, MySQL underlying logic performs a rotate on this no-op
event, but it just doesn't tell Raft about it. If Raft gives up
switching to primary, we're left in a state where Raft works off an
old relay log pointer, which may be invalid.

Registering the pointer here will cause 2 registers during transtion to primary,
one for the no-op rotate, and another when we actually switch to binlog
mode. I feel that's fine, especially since it keeps Raft in the know
about what MySQL is doing to the logs during a transition.

Differential Revision: D57941368

fbshipit-source-id: 6b86761b64333a614c14a3f3f3a290c7dce60ce6