Parsing blob to float* directly when needed
Summary:
When we will not change the input arg, we can cast blob to float* directly to avoid copying. When we will change the input arg, such as in function FB_VECTOR_NORMALIZE_L2, we still need to copy.

The main change is in `Fb_vector` class which consolidates getting information for both json and blob parsed result.

Differential Revision: D58095690

fbshipit-source-id: 1be9017959738d890c89db31e15baef3ab758622