Add support for PK pre-filtering on (VAR)CHARs for vector search
Summary:
If CHAR or VARCHAR key parts are used in the primary key where a vector index is present,
for PK "pre-filtering" to work, the expected KEY/VALUE are needed to unpack the primary key
parts correctly on to table record for condition evaluation.

In the above scenario, the vector index VALUE will look like this:

  [DATA TAG] [VECTOR CODES] [UNPACK INFO FOR CHARS/VARCHARS]

The vector codes are not expected during unpacking and therefore need to be extracted out.

Differential Revision: D58181173

fbshipit-source-id: 67e13f384cbb3baf59082f16ce63723ff57a3f57