Add vector index plan details in EXPLAIN TREE
Summary:
During query planning for vector search, the optimizer can pick various plans. For any of the complex plans involving JOINs, subqueries etc, the plan picked will likely not be a vector index based plan. For such plans, the EXPLAIN output will be sufficient.

When the optimizer picks the vector index, we can differentiate further into the following:
- ordered vector index scan plan (where the vector embeddings are scanned, filtered based on PK key parts, and pushed to FAISS for similarity search)
- vector index scan plan (where vector embeddings are scanned for the nearest nprobe lists, and the corresponding rows read and passed up the SQL execution stack).

To understand which plan got picked, whether on a PROD shard, or while debugging on your test setup, you can check the EXPLAIN FORMAT=TREE output.

For the ordered index scan, the output will look like the following:

  EXPLAIN FORMAT=TREE SELECT *, fb_vector_ip(vector1, '[1, 0.9]') AS dis
  FROM t1 ORDER BY dis DESC LIMIT 10;
  EXPLAIN
  -> Limit: 10 row(s)  (rows=1)
    -> Ordered vector index scan on t1 using key1 (reverse)  (rows=1)

For the vector index scan that relies on filesort for oredering:

  EXPLAIN FORMAT=TREE SELECT *, fb_vector_ip(vector1, '[1, 0.9]') AS dis
  FROM t1 ORDER BY dis DESC LIMIT 10;
  EXPLAIN
  -> Limit: 10 row(s)  (rows=1)
    -> Sort: dis DESC, limit input to 10 row(s) per chunk  (rows=1)
        -> Vector index scan on t1 using key1 (reverse)  (rows=1)

If there are any primary key conditions that can be pushed to the storage engine, they will show up as index conditions:

  EXPLAIN FORMAT=TREE SELECT *, fb_vector_l2(vector1, '[12.5, 3.5, 2.5]') AS dis
  FROM t1
  WHERE
    pk_char_keypart_1 in ('one', 'five') OR
    pk_bigint_keypart_2 IN (2, 6) OR
    pk_bigint_keypart_2 LIKE '1%'
  ORDER BY dis LIMIT 10;

  EXPLAIN
  -> Limit: 10 row(s)  (rows=1)
    -> Filter: ((t1.pk_char_keypart_1 in ('one','five')) or (t1.pk_bigint_keypart_2 in (2,6)) or (t1.pk_bigint_keypart_2 like '1%'))  (rows=1)
        -> Ordered vector index scan on t1 using vector_key_1, with index condition: ((t1.pk_char_keypart_1 in ('one','five')) or (t1.pk_bigint_keypart_2 in (2,6)) or (t1.pk_bigint_keypart_2 like '1%'))  (rows=1)

Differential Revision: D58452680

fbshipit-source-id: 987105a9dfd727216b52a6b3ae3df5e507c9e8dd