Fix flakiness in rpl_raft_shardbeats
Summary:
We recently added returning an error when we're trying to append a write while there is a promotion in progress. This means that if we promote between instances very fast, there is a chance that one of the promotions will fail.

We need loosen up the error checks during promotions but this would be a plugin change. A sleep will do for now until we come up with a better fix on the server.

Differential Revision: D58607281

fbshipit-source-id: 7b75f3fe978a53e9391b493df1cc08c6c49dfe83