Support evermeta in mysqldump --dump-fbobj-assoc-stats
Summary: As a part of the execution for re-attributing everstore cost on UDB to its own platform for MCP we want to start collecting the usage of evermeta similar to how other UDB data demand is collected.  All existing UDB table demanded is collected through daily logical backup run which under the hood use mysqldump --dump-fbobj-assoc-stats command. This change is including `evermeta` table as a new field type (without further breakdown of IDs) in the output for this command.

Differential Revision: D59171858

fbshipit-source-id: c3b0f79ed439a003f26c510f931a6eace115b29a