Disabling read free replication when the server is operating in idempotent mode
Summary:
When operating in idempotent mode on the replica, it's
important the we read the correct version of the row that exists in the
engine, we can't just rely on the before image of the row in the binlog.
This is because the before image and the current version of the row in
the engine might differ when we're running idempotent recovery and if we
just use the before image it might lead to inconsistencies in the PK and
other SK when we do a blind write during read free replication. To solve
this we disable read free replication when we're executing in idempotent
mode.

Differential Revision: D59478062

fbshipit-source-id: ba1efd5407df2f65a22df6e43578819e97736536