Admit statements to admission control
Summary:
Currently, there are queries like create/drop table, which enter the scheduler in the sql prepare phase, and stall the scheduler, without giving us the flexibility of excluding them from admission control and thus, the scheduler.

So, this change allows to admit such queries to scheduler first and then, decide whether it needs to be filtered out providing the flexibility of excluding some known expensive operations like create/drop table which cause stalls later on.

the scheduler checks: thd->lex->sql_command, which should be create/drop table.

Have not added a generic change to execute command yet. We will observe the effects of this change first.

Differential Revision: D59736374

fbshipit-source-id: 6c640249ed47fc31c6e3c8ab81bc4362ab8fcff1