Adding support for database rewrite using session variable
Summary:
To support rollback on resharding, we need to apply binlogs from one database to another. After this diff, we can just run the following code at the start of the binlog application:
```
SET SESSION binlog_db_rewrite_map = 'old_db : new_db';
```
This rename only needs to be handled in 2 places:
1. Table Map Log Event
2. Query Log Event

NOTE: If the customer is using `database_name.table_name` in their raw DDL query, the `database_name` in the query won't be updated. This shouldn't matter for us as we block x-database operations. To make sure this is working as expected, we will add a test to ensure that such a query fails after rewriting.

Differential Revision: D59792716

fbshipit-source-id: 721c200d77c1b0a92a5cc20065435894479e6640