Check rocksdb::Transaction::SetName() output
Summary:
Prior to this diff, when MyRocks calls rocksdb::Transaction::SetName()
and it does not have any name, MyRocks calls SIGABRT to abort the MySQL
instance then returns. However, signal handling is asynchronous, so
there are chances that the transactions with empty name can move forward
the transactions. This is not valid and MyRocks needs to explicitly fail
the transactions.

Differential Revision: D60420876

fbshipit-source-id: f93891ccfffa1ecbae3212d3e1d2abd95bce7456