mysqld checksum: use 0s for last null byte for col list mode
Summary:
In
https://github.com/facebook/mysql-5.6/commit/062f145bf8f54a4c19e0a5203e148e937bfffdc3
the checksum computation for `CHECKSUM TABLE` was fixed by ensuring the
last bits in the last null byte of the record was set to a known, fixed
value. That change fixed the nondeterminism reported in
https://bugs.mysql.com/bug.php?id=13710

With the recent addition of column list support in D56383480, I
uncovered a problem.

The normalization implemented in commit
062f145bf8f54a4c19e0a5203e148e937bfffdc3 chose to set all
otherwise-undefined bits to 1. While this is a known, consistent value,
it is sensitive to the number of nullable columns in the table.

In particular, adding or removing a nullable column changes the
`null_mask` computation. We then `OR` the `null_mask` onto the last null
byte, which changes the bit pattern in this byte, and thus changing the
overall checksum.

This is particularly important in AOSC, which changes schemas routinely.

This change slightly adjusts the logic,

OLD approach (no longer in the latest iteration of this diff):
When column selection is used, to `AND` the `NOT` of the `null_byte`
instead, thereby zeroing out those bits, making the value insensitive to
the number of nullable columns. This means that filtered checksums over
the same columns will yield the same value, instead of spuriously
resulting in different values due to the mere existence of other,
unselected, nullable columns.

Additionally, I am resetting the null bits for columns that are not
included in the column list, as otherwise they would affect the row's
checksum. See the comment for an example.

NEW approach:
The old approach listed above had a few issues:
1. Adding columns in the middle would affect the null bit position of any nullable columns after it, so if they happened to be NULL, the checksum would change, since the overall null bitmap would be different.
2. If the table was altered to add many new nullable columns, it would need more null bytes. Even with the zeroing in the old approach, the result would be different since the number of bytes we're checksumming would differ.

For these reasons, I implemented herman's suggestion to extract the used null bits into their own buffer, and checksum just that. That insulates us from both issues noted above.

Differential Revision: D60490673

fbshipit-source-id: 4d106caad6118a19f84d4b955b2df42281db06a9