mysqld checksum: respect column order in statement
Summary:
With the `CHECKSUM TABLE` extensions, we originally did not respect the
column order given. I.e.

  CHECKSUM TABLE t1 (a, b)

was equivalent to

  CHECKSUM TABLE t1 (b, a)

Because all we would use the item_list for was to set the read_set bits,
but the traversal always happened in the field order on the `TABLE`.

This diff changes that to respect the order provided in the statement.

Differential Revision: D60608640

fbshipit-source-id: 7a2937d31f707b02e0f9e46e2e5f7899b381b3f0