Make MyRocks transactions never expire additional rows with TTL after starting
The TTL row filtering is governed by Rdb_transaction::m_snapshot_timestamp
field. This field sometimes get reset to zero when a snapshot is released,
indicating that it should be reinitialized at the next snapshot, or to the
current timestamp (when a single statement in a multi-statement transaction gets
rolled back). The timestamp going forward means that additional rows may be
hidden in the same transaction if they happened to expire between the two
snapshots. Change it so that additional rows are never expired:

- Only initialize the field once per transaction, when the first snapshot is
  created.
- Do not touch it for single-statement rollbacks in multi-statement
  transactions.
- Rename the field to m_earliest_snapshot_ts to better reflect what it does.
- Change the global transaction list to be sorted by the oldest snapshot
  timestamp, so that the oldest global one can be found by reading the
  transaction at the list head. The transactions without snapshot are inserted
  to the its tail, and reinserted as needed upon acquiring the first snapshot.
  To make reinsertions fast, a third list pointer is maintained that points to
  the newest transaction with a snapshot.
- Extend the SHOW ENGINE ROCKSDB TRANSACTION STATUS output with transactions
  that currently do not have a snapshot, for both cases of having a snapshot
  before and never having had one.
- Add a test to rocksdb.ttl_primary_read_filtering for the behavior
  difference introduced by this commit.

At the same time:
- Initialize m_earliest_snapshot_ts from the binlog HLC instead of the current
  system time.

Note for downstreams: if you don't carry the binlog HLC patch,
binlog_hlc_or_snapshot_ts method simplifies to a snapshot->GetUnixTime call.