Adding a field in metadata event to tell if it is part of a mysql transaction
Summary:
In cases where we strip off gtid event from the transactions
before replaying them on a server we were not able to tell if the
metadata event is part of a trx or if it's a free floating event e.g.
carrying prev hlc or opid for rotate event etc. Adding a field in
metadata event to tell if it is part of a mysql transaction.

Differential Revision: D61991631

fbshipit-source-id: 06729e9a2c5c0ccfa4b48aba7a0295a234d4d3ef