Fixing read filtering ts assigned to consistent snapshots
Summary:
When assigning a read filtering ts for a consistent snapshot we
need to take into account all active trxs and set it to the min read
filtering time across all active trxs instead of using the current HLC.
This is because some of the active trxs can commit to binlog later with
a lower read filtering time than the snapshot time. Later if we use the
snapshot for dump/load and then apply binlogs from the snapshot's binlog
cursor we can run into a situation where a binlog trx with lower read
filtering time will fail to execute because it won't find some rows in
the database (because those rows were expired according to the
snapshot's read filtering time).

Also providing a session var for clients to set a read filtering time
for consistent snapshots manually. This will be useful to compare
snapshots between two different servers that are not replicating from
each other.

Differential Revision: D62511731

fbshipit-source-id: 7611d356096af1586dee17bc5d41c50f3fdc2cf2