Support freeing large writebatches
Summary:
Since writebatches are resized on re-use, it's possible the memory
allocated for the writebatch is not freed. This causes replication
threads to hold on to the memory of large writebatches until the threads
are stopped and the writebatches are destructed.

Add a parameter to enable destroying large writebatches. If the size of
the writebatch is above the parameter, free it instead of saving it for
re-use.

Differential Revision: D63550377

fbshipit-source-id: ee824411f2d5fcc4aebf6531fac8136e3e4a9040