Perform thd_status aggregation using mdl mutex
Summary: We have seen that there is high contention and stalls during status aggregation. The mutex used over there does not work well with scheduler. Rather, we want a mutex which allows the waiters to yield the scheduler when waiting for the lock and get notified of the lock when available.

Differential Revision: D63607467

fbshipit-source-id: 7d5b7e06b344271619522f73aa2d196d0a19ef7a