Create psi thread for server session conditionally
Summary:
With server session registering in PSI, this causes some tests to fail as additional keys/threads appear in the performance schema tables now if the client has already created a PFS thread. The server was probably built with the expectation that only the client invoking its usage is responsible for PSI but we have recently made some changes here, so that it can also register directly, if no THD was supplied in the Srv_session constructor which made session create one of its own.

In this light, altering the code here to only create a new pfs thread when none is already created.

Also, creating a new PSI key to make it easier to distinguish.

Differential Revision: D63711995

fbshipit-source-id: a911cea7a33599775bffdef237e19d14b7beb87e