add error checking for missing unpack info
Summary:
For vector index scans, typically unpack info is not required, as we typically only want the vector codes from the vector index. The unpack info required to decode CHARs, VARCHARs, etc can be obtained from the table row. But for PK filtering using ICP the  unpack info is required for the filtering to work.

We had one version incompatibility issue earlier where the missing unpack info caused a crash in prod. Repeated queries would cause the crashes to repeat. The right way to handle this would be to return an error on detecting the missing unpack info in the field decode path.

This diff returns errors for this case, and adds test cases for two charsets (with two different decode functions).

Differential Revision: D64347750

fbshipit-source-id: 2f2d5d3011490b113d58893c08b3a03fec67e4e9